OPHJOD-814: Populate "Osaamiseni" profile page with "Muu osaaminen"

+++ src/routes/Profile/Competences/Competences.tsx
@@ -1,4 +1,5 @@
 import { client } from '@/api/client';
+import { components } from '@/api/schema';
 import {
   MainLayout,
   RoutesNavigationList,
@@ -7,21 +8,12 @@ import {
   type RoutesNavigationListProps,
 } from '@/components';
 import { OsaaminenLahdeTyyppi } from '@/components/OsaamisSuosittelija/OsaamisSuosittelija';
-import { OSAAMINEN_COLOR_MAP } from '@/constants';
 import { useActionBar } from '@/hooks/useActionBar';
+import { Filters } from '@/routes/Profile/Competences/Filters';
 import { GroupByAlphabet } from '@/routes/Profile/Competences/GroupByAlphabet';
 import { GroupBySource } from '@/routes/Profile/Competences/GroupBySource';
 import { CompetencesLoaderData } from '@/routes/Profile/Competences/loader';
-import {
-  Accordion,
-  Button,
-  Checkbox,
-  Modal,
-  RadioButton,
-  RadioButtonGroup,
-  RoundButton,
-  useMediaQueries,
-} from '@jod/design-system';
+import { Button, Modal, RoundButton, useMediaQueries } from '@jod/design-system';
 import React from 'react';
 import { createPortal } from 'react-dom';
 import { useTranslation } from 'react-i18next';
@@ -30,18 +22,40 @@ import { useLoaderData, useOutletContext } from 'react-router-dom';
 import { mapNavigationRoutes } from '../utils';
 import { FILTERS_ORDER, GROUP_BY_ALPHABET, GROUP_BY_SOURCE, GROUP_BY_THEME, type FiltersType } from './constants';
 
+interface Kokemus {
+  id?: string;
+  uri?: string;
+  nimi: Record<string, string | undefined>;
+  kuvaus?: Record<string, string | undefined>;
+  alkuPvm?: string;
+  loppuPvm?: string;
+  osaamiset?: string[];
+}
+
 const Competences = () => {
   const routes = useOutletContext<RoutesNavigationListProps['routes']>();
-  const { toimenkuvat, koulutukset, patevyydet, osaamiset: osaamisetData } = useLoaderData() as CompetencesLoaderData;
+  const {
+    toimenkuvat,
+    koulutukset,
+    patevyydet,
+    muutOsaamiset,
+    osaamiset: osaamisetData,
+  } = useLoaderData() as CompetencesLoaderData;
   const [initialized, setInitialized] = React.useState(false);
   const [osaamiset, setOsaamiset] = React.useState(osaamisetData);
   const { t, i18n } = useTranslation();
-  const title = t('profile.competences');
+  const title = t('profile.competences.title');
   const [groupBy, setGroupBy] = React.useState<string>(GROUP_BY_SOURCE);
   const navigationRoutes = React.useMemo(() => mapNavigationRoutes(routes), [routes]);
   const actionBar = useActionBar();
   const locale = i18n.language as 'fi' | 'sv';
-  const [selectedFilters, setSelectedFilters] = React.useState<FiltersType>({});
+  const [selectedFilters, setSelectedFilters] = React.useState<FiltersType>({
+    TOIMENKUVA: [],
+    KOULUTUS: [],
+    PATEVYYS: [],
+    MUU_OSAAMINEN: [],
+    KIINNOSTUS: [],
+  });
   const [filterKeys, setFilterKeys] = React.useState<(keyof FiltersType)[]>([]);
   const [showFilters, setShowFilters] = React.useState(false);
   const { sm } = useMediaQueries();
@@ -49,14 +63,14 @@ const Competences = () => {
   const mapExperienceToFilter = React.useCallback(
     (currentFilters: FiltersType, type: OsaaminenLahdeTyyppi) => (experience: Kokemus) => ({
       label: experience.nimi[locale] ?? '',
-      value: experience.id ?? '',
+      value: experience.id ?? experience.uri ?? '',
       checked: currentFilters?.[type]?.find((item) => item.value === experience.id)?.checked ?? true,
     }),
     [locale],
   );
 
   const initFilters = React.useCallback(
-    (osaamiset: OsaaminenApiResponse[]) => {
+    (osaamiset: components['schemas']['YksilonOsaaminenDto'][]) => {
       const initialFilters = {} as FiltersType;
 
       osaamiset.forEach((osaaminen) => {
@@ -72,11 +86,14 @@ const Competences = () => {
         if (type === 'PATEVYYS') {
           initialFilters[type] = patevyydet.map(mapExperienceToFilter(selectedFilters, type));
         }
+        if (type === 'MUU_OSAAMINEN') {
+          initialFilters[type] = muutOsaamiset.map(mapExperienceToFilter(selectedFilters, type));
+        }
       });
 
       return initialFilters;
     },
-    [koulutukset, mapExperienceToFilter, patevyydet, selectedFilters, toimenkuvat],
+    [koulutukset, mapExperienceToFilter, muutOsaamiset, patevyydet, selectedFilters, toimenkuvat],
   );
 
   React.useEffect(() => {
@@ -92,7 +109,11 @@ const Competences = () => {
     }
   }, [initFilters, initialized, osaamiset]);
 
-  const deleteOsaaminen = async (id: string) => {
+  const deleteOsaaminen = async (id?: string) => {
+    if (!id) {
+      return;
+    }
+
     try {
       await client.DELETE('/api/profiili/osaamiset', {
         params: { query: { ids: [id] } },
@@ -108,47 +129,23 @@ const Competences = () => {
     }
   };
 
-  const toggleSingleFilter = (type: OsaaminenLahdeTyyppi, index: number) => () => {
-    const newFilter = { ...selectedFilters };
-    newFilter[type] = newFilter[type] ?? [];
-
-    newFilter[type]![index] = newFilter[type]![index] || {};
-
-    if (newFilter[type]?.[index]) {
-      newFilter[type]![index].checked = !newFilter[type]![index].checked;
-      setSelectedFilters(newFilter);
-    }
-  };
-
-  const isOsaaminenVisible = (type: OsaaminenLahdeTyyppi, id: string): boolean =>
-    selectedFilters[type]?.find((item) => item.value === id)?.checked ?? false;
-
-  const isFilterTypeChecked = (type: OsaaminenLahdeTyyppi) =>
-    (Array.isArray(selectedFilters[type]) &&
-      selectedFilters[type]!.length > 0 &&
-      selectedFilters[type]?.some((item) => item.checked)) ??
-    false;
-
-  const toggleFiltersByType = (type: OsaaminenLahdeTyyppi) => () => {
-    const newFilter = { ...selectedFilters };
-    newFilter[type] = newFilter[type] ?? [];
-
-    // If there are any checked items, uncheck all, otherwise check all
-    const targetState = !isFilterTypeChecked(type);
-
-    newFilter[type]?.forEach((item) => {
-      item.checked = targetState;
-    });
-
-    setSelectedFilters(newFilter);
-  };
+  // Determines if osaamiset from a specific source should be visible. Id is the id of the source (eg. koulutus or toimenkuva).
+  // Muu osaaminen doesn't have any sources, so they'll show up if any are marked as checked.
+  const isOsaaminenVisible = React.useCallback(
+    (type: OsaaminenLahdeTyyppi, id?: string): boolean => {
+      return type === 'MUU_OSAAMINEN'
+        ? selectedFilters.MUU_OSAAMINEN.length > 0 && selectedFilters.MUU_OSAAMINEN.some((item) => item.checked)
+        : selectedFilters[type]?.find((item) => item.value === id)?.checked ?? false;
+    },
+    [selectedFilters],
+  );
 
   const MobileFilterButton = () => {
     return !sm ? (
       <RoundButton
         size="sm"
         bgColor="white"
-        label="Näytä suodattimet"
+        label={t('profile.competences.show-filters')}
         hideLabel
         onClick={() => setShowFilters(true)}
         icon={<MdTune size={24} />}
@@ -158,62 +155,6 @@ const Competences = () => {
     );
   };
 
-  const Filters = () => (
-    <>
-      <SimpleNavigationList title="Järjestele" backgroundClassName={sm ? 'bg-bg-gray-2' : 'bg-bg-gray'} collapsible>
-        <RadioButtonGroup value={groupBy} onChange={setGroupBy} className="py-4" label="Järjestele (kaikki)" hideLabel>
-          <RadioButton label="Lähteiden mukaan" value={GROUP_BY_SOURCE} />
-          <RadioButton label="Teemoittain" value={GROUP_BY_THEME} />
-          <RadioButton label="Aakkosellisesti" value={GROUP_BY_ALPHABET} />
-        </RadioButtonGroup>
-      </SimpleNavigationList>
-      <SimpleNavigationList title="Suodata" backgroundClassName={sm ? 'bg-bg-gray-2' : 'bg-bg-gray'} collapsible>
-        <div className="flex flex-col gap-y-3 py-4">
-          {filterKeys.map((key) => (
-            <Accordion
-              key={key}
-              title={
-                <Checkbox
-                  label={
-                    <span className="flex items-center hyphens-auto" lang={locale}>
-                      <div
-                        className={`mx-3 h-5 w-5 flex-none rounded-full ds-bg-tag-${OSAAMINEN_COLOR_MAP[key]}`}
-                        aria-hidden
-                      />
-                      {t(`types.competence.${key}`)}
-                    </span>
-                  }
-                  checked={isFilterTypeChecked(key)}
-                  onChange={toggleFiltersByType(key)}
-                  ariaLabel="Työpaikka osaamiset"
-                  name="suodata"
-                  value="tyopaikka-osaamiset"
-                  className="min-h-7"
-                />
-              }
-              expandMoreText={t('expand-more')}
-              expandLessText={t('expand-less')}
-              lang={locale}
-            >
-              {selectedFilters[key]?.map((item, idx) => (
-                <div className="pl-6" key={item.value}>
-                  <Checkbox
-                    name={item.label}
-                    ariaLabel={`${key} ${item.label}`}
-                    label={item.label}
-                    checked={item.checked}
-                    onChange={toggleSingleFilter(key, idx)}
-                    value={item.value}
-                  />
-                </div>
-              ))}
-            </Accordion>
-          ))}
-        </div>
-      </SimpleNavigationList>
-    </>
-  );
-
   return (
     <MainLayout
       navChildren={
@@ -221,7 +162,13 @@ const Competences = () => {
           <SimpleNavigationList title={t('profile.index')}>
             <RoutesNavigationList routes={navigationRoutes} />
           </SimpleNavigationList>
-          <Filters />
+          <Filters
+            filterKeys={filterKeys}
+            groupBy={groupBy}
+            selectedFilters={selectedFilters}
+            setGroupBy={setGroupBy}
+            setSelectedFilters={setSelectedFilters}
+          />
         </div>
       }
     >
@@ -236,10 +183,18 @@ const Competences = () => {
           <Modal
             open={showFilters}
             onClose={() => setShowFilters(false)}
-            content={<Filters />}
+            content={
+              <Filters
+                filterKeys={filterKeys}
+                groupBy={groupBy}
+                selectedFilters={selectedFilters}
+                setGroupBy={setGroupBy}
+                setSelectedFilters={setSelectedFilters}
+              />
+            }
             footer={
               <div className="flex flex-row justify-end gap-4">
-                <Button variant="white" label="Sulje" onClick={() => setShowFilters(false)} />
+                <Button variant="white" label={t('close')} onClick={() => setShowFilters(false)} />
               </div>
             }
           />
@@ -274,25 +229,25 @@ const Competences = () => {
             <Button
               disabled
               variant="white"
-              label="Lisää osaaminen"
+              label={t('profile.competences.add')}
               onClick={() => {
-                alert('Lisää');
+                alert(t('profile.competences.add'));
               }}
             />
             <Button
               disabled
               variant="white"
-              label="Jaa"
+              label={t('share')}
               onClick={() => {
-                alert('Jaa');
+                alert(t('share'));
               }}
             />
             <Button
               disabled
               variant="white"
-              label="Muokkaa"
+              label={t('edit')}
               onClick={() => {
-                alert('Muokkaa');
+                alert(t('edit'));
               }}
             />
           </div>,
